WIP moving claims mapping into service
